+++ src/app/routes/connector-ui/connector-ui.module.ts
@@ -17,6 +17,7 @@ import {ConnectorUiComponent} from './connector-ui.component';
 import {ContractAgreementPageModule} from './contract-agreement-page/contract-agreement-page.module';
 import {ContractDefinitionPageModule} from './contract-definition-page/contract-definition-page.module';
 import {DashboardPageModule} from './dashboard-page/dashboard-page.module';
+import {LocationHistoryUtils} from './logout-page/location-history-utils';
 import {LogoutPageModule} from './logout-page/logout-page.module';
 import {PreviousRouteListener} from './logout-page/previous-route-listener';
 import {PolicyDefinitionPageModule} from './policy-definition-page/policy-definition-page.module';
@@ -55,7 +56,7 @@ import {TransferHistoryPageModule} from './transfer-history-page/transfer-histor
     ConnectorUiRoutingModule,
   ],
   declarations: [ConnectorUiComponent],
-  providers: [PreviousRouteListener],
+  providers: [PreviousRouteListener, LocationHistoryUtils],
 })
 export class ConnectorUiModule {
   constructor(previousRouteListener: PreviousRouteListener) {